requirements.txt
Flask==2.0.1
Pillow>=8.3.1
Flask-CORS==3.0.10
rembg
Werkzeug
python-dotenv>=1.0.1